% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_pipe_improved.R
\name{insert_pipe_improved}
\alias{insert_pipe_improved}
\title{Improved Pipe Insertion}
\usage{
insert_pipe_improved()
}
\value{
none
}
\description{
Inserts `%>%` and then starts a new line at the cursor position.
}
