% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_underscore.R
\name{insert_underscore}
\alias{insert_underscore}
\title{Underscore Insertion}
\usage{
insert_underscore()
}
\value{
none
}
\description{
Inserts "_" character. Included for assigning a Shift + Spacebar or Ctrl + Spacebar
shortcut for keyboards that include those keys in a thumb cluster arrangement
}
